[SERV-1877] renaming and test stub
